Support recusive get spark column from expr
