+++ java/openmldb-batch/src/main/scala/com/_4paradigm/openmldb/batch/nodes/JoinPlan.scala
@@ -108,123 +108,9 @@ object JoinPlan {
     // extra conditions
     if (filter.condition() != null) {
       if (ctx.getConf.enableJoinWithNativeExpr) {
-        val expr = filter.condition()
-        expr.GetExprType() match {
-          case ExprType.kExprBinary =>
-            val binaryExpr = BinaryExpr.CastFrom(expr)
-            val op = binaryExpr.GetOp()
-            op match {
-              case FnOperator.kFnOpAnd =>
-                // TODO(tobe): Only support for binary sub expressions
-                val leftExpr = BinaryExpr.CastFrom(binaryExpr.GetChild(0))
-                val rightExpr = BinaryExpr.CastFrom(binaryExpr.GetChild(1))
-
-                val leftCondition = leftExpr.GetOp() match {
-                  case FnOperator.kFnOpNeq =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left.notEqual(right)
-                  case FnOperator.kFnOpLt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left < right
-                  case FnOperator.kFnOpLe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left <= right
-                  case FnOperator.kFnOpGt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left > right
-                  case FnOperator.kFnOpGe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left >= right
-                }
-
-                val rightCondition = rightExpr.GetOp() match {
-                  case FnOperator.kFnOpNeq =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left.notEqual(right)
-                  case FnOperator.kFnOpLt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left < right
-                  case FnOperator.kFnOpLe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left <= right
-                  case FnOperator.kFnOpGt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left > right
-                  case FnOperator.kFnOpGe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left >= right
-                }
-
-                joinConditions += leftCondition.and(rightCondition)
-              case FnOperator.kFnOpOr =>
-
-                // TODO(tobe): Only support for binary sub expressions
-                val leftExpr = BinaryExpr.CastFrom(binaryExpr.GetChild(0))
-                val rightExpr = BinaryExpr.CastFrom(binaryExpr.GetChild(1))
-
-                val leftCondition = leftExpr.GetOp() match {
-                  case FnOperator.kFnOpNeq =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left.notEqual(right)
-                  case FnOperator.kFnOpLt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left < right
-                  case FnOperator.kFnOpLe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left <= right
-                  case FnOperator.kFnOpGt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left > right
-                  case FnOperator.kFnOpGe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(leftExpr, node, leftDf, rightDf)
-                    left >= right
-                }
-
-                val rightCondition = rightExpr.GetOp() match {
-                  case FnOperator.kFnOpNeq =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left.notEqual(right)
-                  case FnOperator.kFnOpLt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left < right
-                  case FnOperator.kFnOpLe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left <= right
-                  case FnOperator.kFnOpGt =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left > right
-                  case FnOperator.kFnOpGe =>
-                    val (left, right) = ExpressionUtil.binaryExprToSparkColumns(rightExpr, node, leftDf, rightDf)
-                    left >= right
-                }
-
-                joinConditions += leftCondition.or(rightCondition)
-
-              case FnOperator.kFnOpNot =>
-              case FnOperator.kFnOpEq =>
-              case FnOperator.kFnOpEq => // Handled by above left_key() and right_key()
-              case FnOperator.kFnOpNeq =>
-                val (left, right) = ExpressionUtil.binaryExprToSparkColumns(binaryExpr, node, leftDf, rightDf)
-                joinConditions += left.notEqual(right)
-              case FnOperator.kFnOpLt =>
-                val (left, right) = ExpressionUtil.binaryExprToSparkColumns(binaryExpr, node, leftDf, rightDf)
-                joinConditions += (left < right)
-              case FnOperator.kFnOpLe =>
-                val (left, right) = ExpressionUtil.binaryExprToSparkColumns(binaryExpr, node, leftDf, rightDf)
-                joinConditions += (left <= right)
-              case FnOperator.kFnOpGt =>
-                val (left, right) = ExpressionUtil.binaryExprToSparkColumns(binaryExpr, node, leftDf, rightDf)
-                joinConditions += (left > right)
-              case FnOperator.kFnOpGe =>
-                val (left, right) = ExpressionUtil.binaryExprToSparkColumns(binaryExpr, node, leftDf, rightDf)
-                joinConditions += (left >= right)
-            }
-          case _ => throw new UnsupportedHybridSeException(
-            s"Join condition does not support expression type ${expr.GetExprType}, try disabling join native expr opt")
-        }
+        joinConditions += ExpressionUtil.recusiveGetSparkColumnFromExpr(filter.condition(), node, leftDf, rightDf)
 
       } else { // Disable join with native expression, use encoder/decoder and jit function
-
         val regName = "SPARKFE_JOIN_CONDITION_" + filter.fn_info().fn_name()
         val conditionUDF = new JoinConditionUDF(
           functionName = filter.fn_info().fn_name(),
