+++ java/openmldb-batch/src/test/scala/com/_4paradigm/openmldb/batch/end2end/unsafe/TestUnsafeJoin.scala
@@ -67,6 +67,8 @@ class TestUnsafeJoin extends SparkTestSuite {
       " ON t1.id >= 1 and t2.id > 1")
     testSql("SELECT t1.id as t1_id, t2.id as t2_id, t1.name FROM t1 LEFT JOIN t2" +
       " ON t1.id >= 1 or t2.id > 1")
+    testSql("SELECT t1.id as t1_id, t2.id as t2_id, t1.name FROM t1 LEFT JOIN t2" +
+      " ON t1.id >= 1 or t2.id > 1 and t1.id = t2.id or t1.id < 10 and t2.id > 0.1 or t2.id = 2")
   }
 
   override def customizedAfter(): Unit = {
